+++ src/Dfe.PlanTech.Application/Questionnaire/Queries/QuestionWithSectionDto.cs
@@ -2,5 +2,4 @@ using Dfe.PlanTech.Domain.Questionnaire.Models;
 
 namespace Dfe.PlanTech.Application.Questionnaire.Queries;
 
-//TODO: Refactor section to own DTO
 public readonly record struct QuestionWithSectionDto(Question Question, string SectionId, string SectionName, string SectionSlug);
